+++ src/Steps/Loading/LoadingStep.php
@@ -9,10 +9,24 @@ abstract class LoadingStep extends Step implements LoadingStepInterface
 {
     protected LoaderInterface $loader;
 
+    protected ?string $useLoaderKey = null;
+
     public function addLoader(LoaderInterface $loader): static
     {
         $this->loader = $loader;
 
         return $this;
+    }
+
+    public function useLoader(string $key): static
+    {
+        $this->useLoaderKey = $key;
+
+        return $this;
+    }
+
+    public function usesLoader(): ?string
+    {
+        return $this->useLoaderKey;
     }
 }