+++ src/Steps/Loading/LoadingStepInterface.php
@@ -8,4 +8,14 @@ use Crwlr\Crawler\Steps\StepInterface;
 interface LoadingStepInterface extends StepInterface
 {
     public function addLoader(LoaderInterface $loader): static;
+
+    /**
+     * Lets the user define which loader instance to use, if the crawler has multiple loaders.
+     */
+    public function useLoader(string $key): static;
+
+    /**
+     * Returns the key of the loader that should be used, if the user defined one, otherwise returns null.
+     */
+    public function usesLoader(): ?string;
 }