+++ tests/Steps/Loading/LoadingStepTest.php
@@ -4,13 +4,15 @@ namespace tests\Steps\Loading;
 
 use Crwlr\Crawler\Input;
 use Crwlr\Crawler\Loader\Http\HttpLoader;
+use Crwlr\Crawler\Logger\CliLogger;
 use Crwlr\Crawler\Steps\Loading\LoadingStep;
+use Crwlr\Crawler\UserAgents\UserAgent;
 use Generator;
 use Mockery;
 
 use function tests\helper_traverseIterable;
 
-test('You can add a loader', function () {
+test('you can add a loader', function () {
     $step = new class () extends LoadingStep {
         protected function invoke(mixed $input): Generator
         {
@@ -27,4 +29,19 @@ test('You can add a loader', function () {
     $step->addLoader($loader);
 
     helper_traverseIterable($step->invokeStep(new Input('https://www.digitalocean.com/blog')));
+});
+
+test('you can set the key of the loader that it should use', function () {
+    $step = new class () extends LoadingStep {
+        protected function invoke(mixed $input): Generator
+        {
+            yield 'yo';
+        }
+    };
+
+    expect($step->usesLoader())->toBeNull();
+
+    $step->useLoader('ftp');
+
+    expect($step->usesLoader())->toBe('ftp');
 });